fix(servicepay): moved logic if payment amount required to viewModel
